packages/compass-client/pyproject.toml
[tool.poetry]
name = "compass-client"
version = "0.1.0"
description = "Compass Education API client library with mock data support"
authors = ["Jethro <jethro@example.com>"]
readme = "README.md"
packages = [{include = "compass_client"}]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
requests = "^2.32.5"
beautifulsoup4 = "^4.12.0"
pydantic = "^2.0.0"
typer = "^0.15.4"
python-dotenv = "^1.0.0"

[tool.poetry.scripts]
compass-client = "compass_client.cli.main:main"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
black = "^23.0.0"
flake8 = "^6.0.0"
mypy = "^1.0.0"
ruff = "^0.14.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"
markers = [
    "integration: marks tests as integration tests requiring real API access (deselect with '-m \"not integration\"')",
]

[tool.flake8]
ignore = ["E501"]
